requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.6
django-allauth==65.4.1
django-crispy-forms==2.3
django-humanize==0.1.2
django-user-visit==2.0
humanize==4.12.2
pillow==11.1.0
sqlparse==0.5.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
